Refactor NeuralNetworkViz and NesEmulator for improved layout and functionality
- Adjusted the dynamic height calculation in NeuralNetworkViz for better visual balance based on the number of layers.
- Enhanced NesEmulator to support a full-width display option, improving the layout in the UI.
- Updated the styling of the canvas in NesEmulator to maintain aspect ratio and improve rendering quality.
- Reorganized the layout in the example page to ensure the emulator and neural network visualization are displayed correctly and responsively.

+++ static/python/ml/neural/neural.py
@@ -33,9 +33,11 @@ class NeuralNetwork:
 
         for i in range(len(layer_sizes) - 1):
             # Xavier initialization for better starting weights
+            # Allows network to learn meaningful associations between inputs and outputs
             w = np.random.randn(layer_sizes[i+1], layer_sizes[i]) * np.sqrt(2.0 / layer_sizes[i])
+            w = w.astype(np.float32)
             # Bias should be 1D array, not 2D column vector
-            b = np.zeros(layer_sizes[i+1])
+            b = np.zeros(layer_sizes[i+1], dtype=np.float32)
 
             self.weights.append(w)
             self.biases.append(b)
